requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==3.2.10
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==11.1.0
feedparser==6.0.8
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
prompt-toolkit==3.0.24
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
